change adc code to handle channels

+++ implementation/adc.h
@@ -4,6 +4,6 @@
 #include <ezdsp5535.h>
 
 void initAdc();
-Int16 readAdcBlocking(Uint8 ch);
+Int16 readAdcBlocking(Uint16 ch);
 
 #endif /*ADC_H_*/
